mlir2chainer: refactoring

+++ mlir2chainer/mlir2chainer/tests/test_bilinear_2d.py
@@ -2,8 +2,6 @@ import mlir
 import chainer
 import mlir2chainer
 import numpy as np
-import six
-import msgpack
 import sys
 import chainer.functions as F
 import util
@@ -21,7 +19,7 @@ def test_bilinear_2d():
     result = mlir.MLIR('Bilinear2D', 'mlir2chainer_test', 0.1, input_names, output_names, nodes, [function])
     result.dump('bilinear_2d.mlir')
     
-    x = np.random.randn(*in_shape)
+    x = np.random.randn(*in_shape).astype(np.float32)
     ref = function.run(x)
     with chainer.using_config('train', False):
         m = mlir2chainer.ChainerNN('bilinear_2d.mlir')
